+++ inc/HM11/hm11.c
@@ -64,8 +64,9 @@ void SysTick_Handler(void) {
     static uint32_t count = 0;
     I2C_system_tick(&g_core_i2c0, 1);
     if(_tx_buffer_ != ( int8_t* ) 0){
-        if (count == 100000){
-            // UART tx every 1000 ticks (1ms)
+        if (count == 5000){
+            // UART tx every 5000 ticks (10ms)
+    		Hm11_Packet();
             UART_send( &g_uart_0, _tx_buffer_, 20 );
             count = 0;
         }